Create modules.py
This is a very basic example and real-world applications would be much more complex. You would also need to implement the frontend of your application, handle error correction, and integrate with any necessary APIs. If you’re new to app development, I would recommend starting with some tutorials or courses to get a feel for it